geek/tos3_migration: add point about disabled services
Disabled transmission is knowns issue caused by rename of package in
upstream. We can't simply fix it because after update of that package we
can't detect that transmission was enabled. This can affect also other
right now unknown services. This could clearly happen to any packages.
There is other possible cause and that is rename of service file itself.

This is generic problem affecting any update. Migration of third-party
services enable is not utilized in upstream as they enable all services
automatically.